+++ src/filter.ts
@@ -36,7 +36,7 @@ export class FieldFilter {
   }
 }
 
-/** Arbitrarily enables some filters (focusing on lossy image comparison). */
+/** Arbitrarily enables some filters. */
 export function enableDefaultFilters(batch: Batch, filters: FieldFilter[]) {
   // Pick the highest (usually meaning slowest) effort for best compression
   // results.
@@ -48,21 +48,6 @@ export function enableDefaultFilters(batch: Batch, filters: FieldFilter[]) {
       const effortFilter = filters[effortIndex];
       effortFilter.enabled = true;
       effortFilter.rangeStart = effortFilter.rangeEnd;
-    }
-  }
-
-  // Restrict JPEG XL to a "reasonable" encoder quality setting range.
-  if (batch.codec.toLowerCase() === 'jxl' ||
-      batch.codec.toLowerCase() === 'jpegxl') {
-    const qualityIndex =
-        batch.fields.findIndex((field: Field) => field.id === FieldId.QUALITY);
-    if (qualityIndex !== -1) {
-      const qualityField = batch.fields[qualityIndex];
-      if (qualityField.isInteger && qualityField.uniqueValuesArray.length > 1) {
-        filters[qualityIndex].enabled = true;
-        filters[qualityIndex].rangeStart =
-            Math.max(filters[qualityIndex].rangeStart, 75);
-      }
     }
   }
 }